Hi!

Thanks for building this and releasing it under MIT licence, It's great to see an implementation of Fortune's algorithm in Rust!

This PR contains some code cleanups and performance enhancements.

Current ``master`` branch:
```
test tests::bench_10000_points ... bench:  81,952,032 ns/iter (+/- 6,801,026)
test tests::bench_100_points   ... bench:     498,563 ns/iter (+/- 8,945)
test tests::bench_1_point      ... bench:       3,233 ns/iter (+/- 183)
```

This pull request:
```
test tests::bench_10000_points ... bench:  23,015,430 ns/iter (+/- 3,501,777)
test tests::bench_100_points   ... bench:     104,735 ns/iter (+/- 1,457)
test tests::bench_1_point      ... bench:       1,138 ns/iter (+/- 25)
```

As you can see, around 3-5x improvement!