@petosegan 2017-11-26
Hey, thanks for these changes! I'm very excited about the performance boosts and the code cleanup is much needed.

I can't merge this yet because it looks like replacing the EventQueue with a BinaryHeap has affected the correctness of the algorithm. I checked this with a toy map generator, which I've belatedly uploaded at https://github.com/petosegan/voronoi_mapgen. It generates garbage after your commit f26632b. If you revert or fix that part then I'll be happy to merge the rest.