@kaedroho 2017-11-26
Thanks for the quick response! I've removed the last three commits (the ID change was a minor performance regression and is only necessary for the BinaryHeap change) so this should be mergable now. I'll look into fixing that part and submit a new pull request when it's ready.