@kaedroho 2017-11-26
I've fixed the issue with ``BinaryHeap`` and added it back to this PR.